requirements.txt
boto3==1.17.101

aws-cdk.core==1.72.0
aws-cdk.aws-lambda==1.72.0
aws_cdk.aws_lambda_event_sources==1.72.0
aws_cdk.aws_lambda_python==1.72.0
aws_cdk.aws_events==1.72.0
aws_cdk.aws_events_targets==1.72.0
aws_cdk.aws_s3==1.72.0
aws-cdk.aws-s3-deployment==1.72.0
